# `hdr_value_at_percentile` optimization.