--- a/src/hdr_histogram.c
+++ b/src/hdr_histogram.c
@@ -165,8 +165,9 @@ static int32_t count_leading_zeros_64(int64_t value)
 #endif
 }
 
-static int64_t get_count_at_index_given_bucket_base_idx(const struct hdr_histogram* h, int32_t bucket_base_idx, int32_t sub_bucket_idx) {
-    return h->counts[bucket_base_idx+sub_bucket_idx-h->sub_bucket_half_count];
+static int64_t get_count_at_index_given_bucket_base_idx(const struct hdr_histogram* h, int32_t bucket_base_idx, int32_t sub_bucket_idx)
+{
+    return h->counts[(bucket_base_idx + sub_bucket_idx) - h->sub_bucket_half_count];
 }
 
 static int32_t get_bucket_base_index(const struct hdr_histogram* h, int32_t bucket_index)
@@ -675,7 +676,8 @@ int64_t hdr_min(const struct hdr_histogram* h)
     return non_zero_min(h);
 }
 
-static int64_t get_value_from_idx_up_to_count(const struct hdr_histogram* h, int64_t count_at_percentile ) {
+static int64_t get_value_from_idx_up_to_count(const struct hdr_histogram* h, int64_t count_at_percentile )
+{
     int64_t count_to_idx = 0;
     int64_t value_from_idx = 0;
     int32_t sub_bucket_idx = -1;
@@ -683,18 +685,20 @@ static int64_t get_value_from_idx_up_to_count(const struct hdr_histogram* h, int
     int32_t bucket_base_idx = get_bucket_base_index(h, bucket_idx);
 
     for (;;) {
-        if (count_to_idx >= count_at_percentile) {
+        if (count_to_idx >= count_at_percentile)
+        {
             break;
         }
         // increment bucket
         sub_bucket_idx++;
-        if (sub_bucket_idx >= h->sub_bucket_count) {
+        if (sub_bucket_idx >= h->sub_bucket_count)
+        {
             sub_bucket_idx = h->sub_bucket_half_count;
             bucket_idx++;
             bucket_base_idx = get_bucket_base_index(h, bucket_idx);
         }
         count_to_idx += get_count_at_index_given_bucket_base_idx(h, bucket_base_idx, sub_bucket_idx);
-        value_from_idx = ((int64_t)(sub_bucket_idx)) << (((int64_t)(bucket_idx))+h->unit_magnitude);
+        value_from_idx = ((int64_t)(sub_bucket_idx)) << (((int64_t)(bucket_idx)) + h->unit_magnitude);
     }
     return value_from_idx;
 }
@@ -705,7 +709,8 @@ int64_t hdr_value_at_percentile(const struct hdr_histogram* h, double percentile
     int64_t count_at_percentile =
         (int64_t) (((requested_percentile / 100) * h->total_count) + 0.5);
     int64_t value_from_idx = get_value_from_idx_up_to_count(h, count_at_percentile);
-    if (percentile == 0.0) {
+    if (percentile == 0.0)
+    {
         return lowest_equivalent_value(h, value_from_idx);
     }
     return highest_equivalent_value(h, value_from_idx);