--- a/test/hdr_histogram_benchmark.cpp
+++ b/test/hdr_histogram_benchmark.cpp
@@ -1,5 +1,5 @@
 #include <benchmark/benchmark.h>
-#include <hdr_histogram.h>
+#include <hdr/hdr_histogram.h>
 #include <math.h>
 #include <random>