This is basically a C port of the optimization in PR https://github.com/HdrHistogram/hdrhistogram-go/pull/48 by @filipecosta90.

before:
```
yoav@yoav-thinkpad:~/HdrHistogram_c/build$ ./test/hdr_histogram_benchmark --benchmark_filter=BM_hdr_value_at_percentile/3/86400000 --benchmark_min_time=10
2022-03-21 12:41:46
Running ./test/hdr_histogram_benchmark
Run on (12 X 4600 MHz CPU s)
CPU Caches:
  L1 Data 32K (x6)
  L1 Instruction 32K (x6)
  L2 Unified 256K (x6)
  L3 Unified 12288K (x1)
Load Average: 0.68, 0.81, 0.89
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
***WARNING*** Library was built as DEBUG. Timings may be affected.
-----------------------------------------------------------------------------------
Benchmark                                         Time             CPU   Iterations
-----------------------------------------------------------------------------------
BM_hdr_value_at_percentile/3/86400000         34937 ns        34955 ns       404859
BM_hdr_value_at_percentile/3/86400000000      36561 ns        36581 ns       398987
```
after:
```
yoav@yoav-thinkpad:~/HdrHistogram_c/build$ ./test/hdr_histogram_benchmark --benchmark_filter=BM_hdr_value_at_percentile/3/86400000 --benchmark_min_time=10
2022-03-21 13:36:17
Running ./test/hdr_histogram_benchmark
Run on (12 X 4600 MHz CPU s)
CPU Caches:
  L1 Data 32K (x6)
  L1 Instruction 32K (x6)
  L2 Unified 256K (x6)
  L3 Unified 12288K (x1)
Load Average: 0.86, 0.96, 1.26
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
***WARNING*** Library was built as DEBUG. Timings may be affected.
-----------------------------------------------------------------------------------
Benchmark                                         Time             CPU   Iterations
-----------------------------------------------------------------------------------
BM_hdr_value_at_percentile/3/86400000          5070 ns         5083 ns      2795009
BM_hdr_value_at_percentile/3/86400000000       5124 ns         5135 ns      2753732
```