@filipecosta90 2022-03-21
great improvement @yoav-steinberg! thank you! 
@mikeb01 do you think this PR can be reviewed? In a nutshell, it avoids wasteful computation on the iterator and just computes the bare minimum up until we reach the value at percentile.