@mikeb01 2022-03-23
This definitely looks interesting.  Which build configuration was used to generate the benchmarks Release or Debug.  If this was Debug, it would be good to rerun them with a release build.