@yoav-steinberg 2022-03-24
> Which build configuration was used...

I used _Release_ for the project itself: 
`cmake -DHDR_HISTOGRAM_BUILD_BENCHMARK=ON -DCMAKE_BUILD_TYPE=Release ..`
but couldn't get the benchmark lib to be built with Release, hence the:
`***WARNING*** Library was built as DEBUG. Timings may be affected.`
But I think it's still valid.