@yoav-steinberg 2022-05-18
@mikeb01 @filipecosta90 I refactored and added an explicit `counts` bounds check as requested for extra safety. Note that added a slight performance hit, so I optimized a bit more to get back to the previous performance. I'd appreciate you verify that I didn't break the logic.
Thanks.