@mikeb01 2022-05-20
I've merged this, but rewrote the loop to match the implementation in the Java version.  The performance measured on my local machine was within 1%.